Tharushi/LegalVision-main/legal_deed_summarizer_flask/requirements.txt
Flask==3.0.3
flask-cors==5.0.0
gunicorn==22.0.0

torch>=2.1.0
transformers>=4.40.0
numpy>=1.24.0
pandas>=2.0.0

evaluate>=0.4.2
rouge_score>=0.1.2

python-dotenv>=1.0.1

spacy>=3.7.5
dateparser>=1.2.0
python-dateutil>=2.9.0.post0


google-genai>=1.0.0
pillow>=10.0.0
python-dotenv>=1.0.1